fix: use navLinks insteand of a tags
