+++ pokeapi/src/Layout.css
@@ -1,4 +1,4 @@
-#root {
+.page {
   display: grid;
   height: 100vh;
   grid-template-columns: 100vw;
