+++ pokeapi/src/Organisms/Header/Navbar.tsx
@@ -3,14 +3,25 @@ import HeaderLinks from "../../Molecules/HeaderLinks/HeaderLinks";
 import type { Link } from "../../Types/Link";
 import "./Navbar.css";
 
-type CustomProps = {
-  links: Link[];
-};
+const links: Link[] = [
+  {
+    name: "Games",
+    link: "games",
+  },
+  {
+    name: "Langs",
+    link: "langs",
+  },
+  {
+    name: "Pokedex",
+    link: "pokedex",
+  },
+];
 
-type Props = CustomProps & React.ButtonHTMLAttributes<HTMLDivElement>;
+type Props = React.ButtonHTMLAttributes<HTMLDivElement>;
 
 export default function Navbar(props: Props) {
-  const { links, className, ...ElementProps } = props;
+  const { className, ...ElementProps } = props;
   const classname = className ?? "";
   return (
     <div className={`${classname} navbar`} {...ElementProps}>
