+++ pokeapi/src/Router.tsx
@@ -3,23 +3,29 @@ import PokemonGame from "./Pages/PokemonGame";
 import PokemonListPage from "./Pages/PokemonListPage";
 import { createBrowserRouter, RouterProvider } from "react-router";
 import ErrorScreen from "./Organisms/ErrorScreen/ErrorScreen";
+import HomePage from "./Pages/HomePage";
 
 const router = createBrowserRouter([
   {
     path: "/",
+    element: <HomePage />,
     children: [
       {
+        index: true,
         path: "/pokedex",
-        Component: PokemonListPage,
+        element: <PokemonListPage />,
       },
       {
-        path: "/games",
-        Component: PokemonGame,
+        path: "games",
+        element: <PokemonGame />,
+      },
+      {
+        path: "*",
+        element: (
+          <ErrorScreen fullScreen={true} error="this page doesnt exist" />
+        ),
       },
     ],
-    errorElement: (
-      <ErrorScreen fullScreen={true} error="this page doesnt exist" />
-    ),
   },
 ]);
 export default function Router(): ReactElement {
